Widen the replay inter-order step counter from Uint16 to Uint32
ReplayWriter::stepsSinceLastOrder was a Uint16 moved over the wire with
writeUint16, so after 65535 order-less ticks (~44 minutes at 25 Hz —
routine in sparse AI runs and 90000-tick trainer games) the counter
silently wrapped, corrupting the tick-to-order alignment of every
subsequent order on read-back.

The counter is now Uint32 in the writer field and at all three write
sites, and VERSION_MINOR is bumped to 87 to version the layout change.
Replays carry no dedicated format version — ReplayReader::loadReplay
checked the embedded VERSION_MAJOR/VERSION_MINOR for exact equality —
so the reader now accepts a range instead: replays from 86 (the
previous format, still Uint16 counters) through the current version,
reading the counter at the width matching the replay's own version.
Older replays and replays from newer builds are rejected as before.

New standalone harness test/ReplayStepCounterTest covers a round-trip
with >65535 quiet steps between orders (which would have wrapped), the
version-86 Uint16 wire layout through the gated read, and the version
floor/ceiling. GameGUI and the Order hierarchy are stubbed following
the NetSendOrderDecodeTest pattern.

+++ src/ReplayWriter.cpp
@@ -102,7 +102,7 @@ void ReplayWriter::pushOrder(std::shared_ptr<Order> order)
 	if (order->getOrderType() == ORDER_VOICE_DATA || order->getOrderType() == ORDER_NULL) return;
 
 	// Write the number of steps since last order to this order (can be 0)
-	buffer->writeUint16(stepsSinceLastOrder, "replayStepsSinceLastOrder");
+	buffer->writeUint32(stepsSinceLastOrder, "replayStepsSinceLastOrder");
 
 	// Write the Order to the file
 	writeOrder(buffer, order, checksum);
@@ -118,7 +118,7 @@ void ReplayWriter::finish()
 	if (!isValid()) return;
 
 	// Write the number of steps since last order to the end of the replay
-	buffer->writeUint16(stepsSinceLastOrder, "replayStepsSinceLastOrder");
+	buffer->writeUint32(stepsSinceLastOrder, "replayStepsSinceLastOrder");
 
 	// We write a NullOrder to mark the end of the replay (like terminating a string with \0)
 	writeOrder(buffer, std::shared_ptr<Order>(new NullOrder()), 0);
@@ -159,7 +159,7 @@ bool ReplayWriter::write(const std::string &filename) const
 	}
 
 	// Write the number of steps since last order to the end of the replay
-	file->writeUint16(0, "replayStepsSinceLastOrder");
+	file->writeUint32(0, "replayStepsSinceLastOrder");
 
 	// Write a NullOrder to the file to make sure it's a NullOrder-terminated replay
 	writeOrder(file, std::shared_ptr<Order>(new NullOrder()), 0);
